make book update logic reusable
