+++ lib/proxy-ws.js
@@ -9,6 +9,7 @@ const Wallet = require('./wallet.js')
 const Orders = require('./orders.js')
 const Positions = require('./positions.js')
 
+const BookScheduler = require('./bookscheduler.js')
 const Book = require('./book.js')
 const formatOrder = require('./hive-order-helper.js')
 
@@ -36,58 +37,24 @@ class ProxyWs extends BaseWs {
 
     this.clients = {}
 
-    this.scheduler = {}
-    this.updateBooks()
-  }
-
-  async getBook (symbol) {
-    const reqId = uuid()
-
-    let res = null
-    const payload = ['get_book_depth', symbol, reqId]
-    try {
-      res = await this.sock.send(reqId, 'gateway', payload)
-    } catch (e) {
-      console.error(e)
-      return
-    }
-
-    return res[0]
-  }
-
-  schedule (id, fun, args, cb) {
-    this.scheduler[id] = setTimeout(async () => {
-      fun.call(fun, args, (err, res) => {
-        if (err) console.error(err)
-        if (res) cb(null, res)
-
-        this.schedule(id, fun, args, cb)
-      })
-    }, 1000)
-  }
-
-  updateBooks () {
-    this.intervalsBook = {}
-
-    this.pairs.forEach((p) => {
-      const task = (p, cb) => {
-        this.getBook(p)
-          .then((res) => { cb(null, res) })
-          .catch((err) => { cb(err) })
-      }
-
-      this.schedule(p, task, p, (err, fullBook) => {
-        if (err) return
-
-        const wsChannel = this.channels.book[p]
-        const book = wsChannel.book
-
-        const msgs = book.update(fullBook)
-        if (msgs.length === 0) return
-
-        this.sendOrderbookUpdates(wsChannel, msgs)
-      })
+    this.bookScheduler = new BookScheduler({
+      pairs: this.pairs,
+      socket: this.sock
     })
+
+    this.bookScheduler.updateBooks(this.handleBookUpdate.bind(this))
+  }
+
+  handleBookUpdate (_err, res) {
+    const [pair, fullBook] = res
+
+    const wsChannel = this.channels.book[pair]
+    const book = wsChannel.book
+
+    const msgs = book.update(fullBook)
+    if (msgs.length === 0) return
+
+    this.sendOrderbookUpdates(wsChannel, msgs)
   }
 
   getBookChannels () {
